Convert content_types to a JS app bundle
