+++ CMakeLists.txt
@@ -669,8 +669,9 @@ if(BUILD_TESTS)
 
   add_e2e_test(
     NAME content_types_test
-    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/content_types.py
+    PYTHON_SCRIPT ${CMAKE_SOURCE_DIR}/tests/js-content-types/content_types.py
     CONSENSUS cft
+    ADDITIONAL_ARGS --js-app-bundle ${CMAKE_SOURCE_DIR}/tests/js-content-types
   )
 
   if(QUOTES_ENABLED)
