+++ tests/js-content-types/app.json
@@ -1,0 +1,52 @@
+{
+  "endpoints": {
+    "/text": {
+      "post": {
+        "js_module": "content_types.js",
+        "js_function": "text",
+        "forwarding_required": "always",
+        "execute_locally": false,
+        "require_client_signature": false,
+        "require_client_identity": true,
+        "readonly": true,
+        "openapi": {}
+      }
+    },
+    "/json": {
+      "post": {
+        "js_module": "content_types.js",
+        "js_function": "json",
+        "forwarding_required": "always",
+        "execute_locally": false,
+        "require_client_signature": false,
+        "require_client_identity": true,
+        "readonly": true,
+        "openapi": {}
+      }
+    },
+    "/binary": {
+      "post": {
+        "js_module": "content_types.js",
+        "js_function": "binary",
+        "forwarding_required": "always",
+        "execute_locally": false,
+        "require_client_signature": false,
+        "require_client_identity": true,
+        "readonly": true,
+        "openapi": {}
+      }
+    },
+    "/custom": {
+      "post": {
+        "js_module": "content_types.js",
+        "js_function": "custom",
+        "forwarding_required": "always",
+        "execute_locally": false,
+        "require_client_signature": false,
+        "require_client_identity": true,
+        "readonly": true,
+        "openapi": {}
+      }
+    }
+  }
+}
